separating app utils from app data utils and moving convert to pandas
